pyproject.toml
[project]
name = "doe-master"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "pillow",
    "matplotlib",
    "opencv-python",
    "scipy",
    "scikit-image",
    "tensorboard",
    "pwlf",
    "pyyaml",
    "easydict",
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]